backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database (updated for Python 3.13 support)
sqlalchemy==2.0.36
psycopg[binary]==3.2.3  # Using psycopg3 (has Python 3.13 support)
alembic==1.13.3

# Data validation (updated for Python 3.13 support)
pydantic==2.10.6
pydantic-settings==2.11.0

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality
black==23.11.0
flake8==6.1.0
